Move sharedModule into sharedFolder/components-directives
