+++ src/app/home/home.module.ts
@@ -8,7 +8,7 @@ import { HomePage } from './home.page';
 import { UpcomingEventsComponent } from './upcoming-events/upcoming-events.component';
 import { EventsTypeListComponent } from './events-type-list/events-type-list.component';
 import { FooterComponent } from './footer/footer.component';
-import { SharedModule } from '../shared-module/shared-module.module';
+import { SharedModule } from 'src/shared/components-directives/shared-module.module';
 
 @NgModule({
   imports: [
