+++ src/shared/components-directives/like-icon/like-icon.component.scss
