+++ src/shared/components-directives/shared-module.module.ts
@@ -1,11 +1,11 @@
-import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { ExploreContainerComponentModule } from '../explore-container/explore-container.module';
-import { FormsModule } from '@angular/forms';
-import { IonicModule } from '@ionic/angular';
 import { LikeIconComponent } from './like-icon/like-icon.component';
 import { TitleFadesComponent } from './title-fades/title-fades.component';
 import { HeaderComponent } from './header/header.component';
+import { CommonModule } from '@angular/common';
+import { ExploreContainerComponentModule } from 'src/app/explore-container/explore-container.module';
+import { FormsModule } from '@angular/forms';
+import { IonicModule } from '@ionic/angular';
+import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
 
 @NgModule({
   imports: [
