+++ src/shared/components-directives/title-fades/title-fades.component.scss
